Parse only one step at a time
Previously, `parseSteps` operated on an array, so there was a lot of
“extra” code to put standalone tooltip data into an array and then
extract just the first element from the result.  This refactors it
to only take a single object at a time.  This should give some (slightly)
improved performance, but is also just easier to understand.

The eventual goal is to make `parseStep` an internal method, and not
require consumers to call it at all.