recalc top/bottom position if tooltip doesn’t fit
Small optimization and usually an edge case, but in the case that the
desired top/bottom position can’t fit the tooltip, there should be a
recalculation. This _still_ doesn’t account for everything (i.e. an
element in the corner) but this is a step in the right direction