Trigger callback when target is not found
This change allows the application to decide what to do when a target
element is not found, rather than the previous behavior of simply 
incrementing/decrementing the index to skip the step.

This way, an app can still accomplish the same behavior by listening to
the callback and incrementing/decrementing the `stepIndex` prop as
appropriate.  The demo has been updated to demonstrate exactly this.

The app could also stop the joyride and then retry after a period of time,
in case it is waiting for an element to be added to its DOM.